+++ src/App.js
@@ -1,27 +1,15 @@
 import React from 'react';
 import Read from './Read.js';
-import { GoogleLogin } from '@react-oauth/google';
-// import SignIn from './SignIn.js';
+import SignIn from './SignIn.js';
 
 function App() {
-  const responseMessage = (response) => {
-    console.log(response);
-};
-const errorMessage = (error) => {
-    console.log(error);
-};
   return (
     <>
-    <h2>React Google Login</h2>
-            <br />
-            <br />
-            <GoogleLogin onSuccess={responseMessage} onError={errorMessage} />
-    {/* <SignIn/> */}
+    <SignIn/>
     <Read/>
     </>
-  );
+  )
 }
-
 export default App;
 
 /*
