+++ src/SignIn.js
@@ -1,13 +1,20 @@
 import React from 'react'; 
+import { GoogleLogin } from '@react-oauth/google';
 
 // https://developers.google.com/identity/sign-in/web/sign-in
 // https://blog.logrocket.com/guide-adding-google-login-react-app/
 
 
 export default function SignIn (){
-    return (
-    <div className="g-signin2" data-onsuccess="onSignIn">
-        <button>Sign In with Google</button>
-    </div>
-    )
+    const responseMessage = (response) => {
+        console.log(response);
+    };
+    const errorMessage = (error) => {
+        console.log(error);
+    };
+      return (
+        <GoogleLogin onSuccess={responseMessage} onError={errorMessage} />
+      );
 }
+
+
